rename bucket reducer
