+++ src/store/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
 
-import items from "reducers/bucket";
+import items from "reducers/cart";
 import category from "reducers/filter";
 
 const reducer = combineReducers({
